requirements.txt
Django==5.1.2
pillow==10.4.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2
psycopg2==2.9.10
django-environ==0.11.2